src/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
